Fix bug that occasionally caused a truncated last letter in track artist
Debugging showed that the "Text" field of Anaconda's Tweet struct
sometimes contains a truncated version of the original tweet message,
which misses the last letter. Consequently, the TrackRecord's "Artist"
field gets initialized incorrectly.
An easy fix for this issue is the use of the "FullText" field instead of
"Text", which seems to contain the correct original tweet message.