Fix bug that caused wrong Kronehit track dates when fetched over midnight
The Kronehit Crawler is able to determine the correct track dates if a
batch of items spans over midnight/two days. However, if the fetch time
is of day B, but the results are only of day A, the track dates are set
to day B. This causes the tracks to be younger than the fetch time and
therefore be skipped. Since the tracks with the incorrect date have
already been fetched correctly by a previous batch iteration, this
behaviour is not affecting the integrity of the collected data, but
neither it is the desired behaviour.