Fix redundant fetches by overlapping tracks with the most recent one
The Kronehit API is not always consistent with its airtimes. A track that gets played at 23:55:23 is processed by the crawler, but its airtime might have changed until the next fetch. This could be caused by certain events, i.e. news reports lasting longer than anticipated and therefore influencing the actual airtime. Kronehit seems to be updating the airtimes to their actual airtime (which definitly makes sense).
Although we are only fetching songs that lay in the past, it still occurs that airtimes changes. To prevent infinite fetches, an overlapping between the most recent track and the oldest track in the batch is created.